# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalPlayers: number;
  totalScore: number;
  activeEvents: number;
  tournamentPlayers: number;
  botOnline: boolean;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  /** @nullable */
  name: string | null;
  score: number;
}

export interface EventItem {
  title: string;
  prize: string;
  playerCount: number;
  startTime: number;
  endTime: number;
  active: boolean;
}

export interface EventsData {
  giveaway: EventItem | null;
  premium: EventItem | null;
}

export interface TournamentData {
  playerCount: number;
  playerIds: string[];
  winners: string[];
}

export type StopRequestType = typeof StopRequestType[keyof typeof StopRequestType];


export const StopRequestType = {
  giveaway: 'giveaway',
  premium: 'premium',
  tournament: 'tournament',
  boss_raid: 'boss_raid',
  taprace: 'taprace',
  marathon: 'marathon',
  all: 'all',
} as const;

export interface StopRequest {
  type: StopRequestType;
}

export interface StopResponse {
  ok: boolean;
}

export type LiveStateTapRaceTaps = {[key: string]: number};

export interface LiveState {
  tapRaceActive: boolean;
  tapRaceStarted: boolean;
  tapRaceMatch: (number | null)[];
  tapRaceTaps: LiveStateTapRaceTaps;
  bossRaidActive: boolean;
  bossRaidStarted: boolean;
  bossRaidHp: number;
  bossRaidMaxHp: number;
  marathonActive: boolean;
  botOnline: boolean;
  /** @nullable */
  lastUpdated: number | null;
}

export type PlaySessionInputType = typeof PlaySessionInputType[keyof typeof PlaySessionInputType];


export const PlaySessionInputType = {
  tap: 'tap',
  raid: 'raid',
} as const;

export interface PlaySessionInput {
  type: PlaySessionInputType;
  chatId: number;
  lobbyDurationMs?: number;
  playDurationMs?: number;
  bossHp?: number;
}

export interface PlaySessionCreated {
  sessionId: string;
  url: string;
}

export interface PlayPlayer {
  playerId: string;
  name: string;
  score: number;
}

export type PlaySessionStateType = typeof PlaySessionStateType[keyof typeof PlaySessionStateType];


export const PlaySessionStateType = {
  tap: 'tap',
  raid: 'raid',
} as const;

export type PlaySessionStateStatus = typeof PlaySessionStateStatus[keyof typeof PlaySessionStateStatus];


export const PlaySessionStateStatus = {
  lobby: 'lobby',
  running: 'running',
  finished: 'finished',
} as const;

export interface PlaySessionState {
  sessionId: string;
  type: PlaySessionStateType;
  status: PlaySessionStateStatus;
  /** @nullable */
  startsAt: number | null;
  /** @nullable */
  endsAt: number | null;
  serverTime: number;
  players: PlayPlayer[];
  /** @nullable */
  bossHp: number | null;
  /** @nullable */
  bossMaxHp: number | null;
  /** @nullable */
  winnerName: string | null;
}

export interface PlayJoinInput {
  /**
     * @minLength 1
     * @maxLength 32
     */
  name: string;
}

export interface PlayJoinResponse {
  playerId: string;
  token: string;
}

export interface PlayActionInput {
  playerId: string;
  token: string;
  /**
     * @minimum 1
     * @maximum 25
     */
  amount: number;
}

export interface PlayActionResponse {
  score: number;
  /** @nullable */
  bossHp: number | null;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};