# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Bot admin dashboard
  - name: play
    description: Web-based mini-games (tap race, boss raid)
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get overall bot stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/leaderboard:
    get:
      operationId: getDashboardLeaderboard
      tags: [dashboard]
      summary: Get ranked leaderboard
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  /dashboard/events:
    get:
      operationId: getDashboardEvents
      tags: [dashboard]
      summary: Get active events
      responses:
        "200":
          description: Active events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsData"

  /dashboard/tournament:
    get:
      operationId: getDashboardTournament
      tags: [dashboard]
      summary: Get tournament state
      responses:
        "200":
          description: Tournament data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentData"

  /dashboard/live:
    get:
      operationId: getDashboardLive
      tags: [dashboard]
      summary: Get live game state
      responses:
        "200":
          description: Live state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiveState"

  /dashboard/stop:
    post:
      operationId: postDashboardStop
      tags: [dashboard]
      summary: Force-stop a giveaway/event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopRequest"
      responses:
        "200":
          description: Stop command queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopResponse"

  /play/session:
    post:
      operationId: createPlaySession
      tags: [play]
      summary: Create a new web-game session (called by bot)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaySessionInput"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaySessionCreated"

  /play/session/{id}:
    get:
      operationId: getPlaySession
      tags: [play]
      summary: Get current session state
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaySessionState"
        "404":
          description: Not found

  /play/session/{id}/join:
    post:
      operationId: joinPlaySession
      tags: [play]
      summary: Register a player for the session
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayJoinInput"
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayJoinResponse"

  /play/session/{id}/action:
    post:
      operationId: actPlaySession
      tags: [play]
      summary: Submit a batch of taps / damage for a player
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayActionInput"
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayActionResponse"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    DashboardSummary:
      type: object
      properties:
        totalPlayers:
          type: integer
        totalScore:
          type: integer
        activeEvents:
          type: integer
        tournamentPlayers:
          type: integer
        botOnline:
          type: boolean
      required:
        - totalPlayers
        - totalScore
        - activeEvents
        - tournamentPlayers
        - botOnline

    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: string
        name:
          type: ["string", "null"]
        score:
          type: integer
      required:
        - rank
        - userId
        - name
        - score

    EventItem:
      type: object
      properties:
        title:
          type: string
        prize:
          type: string
        playerCount:
          type: integer
        startTime:
          type: number
        endTime:
          type: number
        active:
          type: boolean
      required:
        - title
        - prize
        - playerCount
        - startTime
        - endTime
        - active

    EventsData:
      type: object
      properties:
        giveaway:
          oneOf:
            - $ref: "#/components/schemas/EventItem"
            - type: "null"
        premium:
          oneOf:
            - $ref: "#/components/schemas/EventItem"
            - type: "null"
      required:
        - giveaway
        - premium

    TournamentData:
      type: object
      properties:
        playerCount:
          type: integer
        playerIds:
          type: array
          items:
            type: string
        winners:
          type: array
          items:
            type: string
      required:
        - playerCount
        - playerIds
        - winners

    StopRequest:
      type: object
      properties:
        type:
          type: string
          enum: [giveaway, premium, tournament, boss_raid, taprace, marathon, all]
      required:
        - type

    StopResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    LiveState:
      type: object
      properties:
        tapRaceActive:
          type: boolean
        tapRaceStarted:
          type: boolean
        tapRaceMatch:
          type: array
          items:
            type: ["integer", "null"]
        tapRaceTaps:
          type: object
          additionalProperties:
            type: integer
        bossRaidActive:
          type: boolean
        bossRaidStarted:
          type: boolean
        bossRaidHp:
          type: integer
        bossRaidMaxHp:
          type: integer
        marathonActive:
          type: boolean
        botOnline:
          type: boolean
        lastUpdated:
          type: ["number", "null"]
      required:
        - tapRaceActive
        - tapRaceStarted
        - tapRaceMatch
        - tapRaceTaps
        - bossRaidActive
        - bossRaidStarted
        - bossRaidHp
        - bossRaidMaxHp
        - marathonActive
        - botOnline
        - lastUpdated

    PlaySessionInput:
      type: object
      properties:
        type:
          type: string
          enum: [tap, raid]
        chatId:
          type: number
        lobbyDurationMs:
          type: integer
        playDurationMs:
          type: integer
        bossHp:
          type: integer
      required:
        - type
        - chatId

    PlaySessionCreated:
      type: object
      properties:
        sessionId:
          type: string
        url:
          type: string
      required:
        - sessionId
        - url

    PlayPlayer:
      type: object
      properties:
        playerId:
          type: string
        name:
          type: string
        score:
          type: integer
      required:
        - playerId
        - name
        - score

    PlaySessionState:
      type: object
      properties:
        sessionId:
          type: string
        type:
          type: string
          enum: [tap, raid]
        status:
          type: string
          enum: [lobby, running, finished]
        startsAt:
          type: ["number", "null"]
        endsAt:
          type: ["number", "null"]
        serverTime:
          type: number
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayPlayer"
        bossHp:
          type: ["integer", "null"]
        bossMaxHp:
          type: ["integer", "null"]
        winnerName:
          type: ["string", "null"]
      required:
        - sessionId
        - type
        - status
        - startsAt
        - endsAt
        - serverTime
        - players
        - bossHp
        - bossMaxHp
        - winnerName

    PlayJoinInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32
      required:
        - name

    PlayJoinResponse:
      type: object
      properties:
        playerId:
          type: string
        token:
          type: string
      required:
        - playerId
        - token

    PlayActionInput:
      type: object
      properties:
        playerId:
          type: string
        token:
          type: string
        amount:
          type: integer
          minimum: 1
          maximum: 25
      required:
        - playerId
        - token
        - amount

    PlayActionResponse:
      type: object
      properties:
        score:
          type: integer
        bossHp:
          type: ["integer", "null"]
      required:
        - score
        - bossHp